Add test suite for SearchBar component
Add named export to the SearchBar component to allow it to be tested in
isolation (unconnected).

Add `global.location` to `test/init.js` so that the SearchBar can access
it in tests.

SearchBar is not currently in use, however having tests is good
preparation for implementing it in the Blockstack Browser.