Add new test suite for SocialAccountItem
Include a new test for the action taken when a visitor clicks on the
"verified check mark", but skip this test for now, as it requires making
one of the util methods it depends on into a class method of the
component. This can be done simply by wrapping it in a class method.